Correct typing error

+++ cleaning.py
@@ -71,7 +71,7 @@ def read_clean_data(path: str = './application_train.csv') -> DataFrame:
 
     # Invert negations and correct data types
     data.DAYS_BIRTH = (-data.DAYS_BIRTH).astype(np.uint16)
-    data.DAYS_EMPLOYED = (-data.DAYS_EMPLOYED).astype(np.uint16)
+    data.DAYS_EMPLOYED = (-data.DAYS_EMPLOYED).astype(np.uint32)
     data.DAYS_REGISTRATION = (-data.DAYS_REGISTRATION).astype(np.uint16)
     data.DAYS_ID_PUBLISH = (-data.DAYS_ID_PUBLISH).astype(np.uint16)
     data.DAYS_LAST_PHONE_CHANGE = (-data.DAYS_LAST_PHONE_CHANGE)
